+++ src/main/java/com/codurance/allaboard/config/security/GoogleTokenAuthenticator.java
@@ -1,4 +1,4 @@
-package com.codurance.allaboard.config;
+package com.codurance.allaboard.config.security;
 
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdTokenVerifier;
 import com.google.api.client.http.javanet.NetHttpTransport;
